requirements.txt
fastapi
uvicorn[standard]
asyncpg
httpx
google-generativeai
openai
pydantic
python-dotenv
apscheduler
